+++ .changeset/tiny-paws-happen.md
@@ -1,0 +1,8 @@
+---
+'@graphql-mesh/odata': patch
+'@graphql-mesh/fusion-composition': patch
+'@omnigraph/odata': patch
+'@graphql-mesh/transport-odata': patch
+---
+
+New Transport and Loader implementation for OData
