+++ e2e/odata-trippin/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@e2e/odata-trippin-example",
+  "license": "MIT",
+  "private": true,
+  "dependencies": {
+    "@graphql-mesh/compose-cli": "workspace:*",
+    "@graphql-mesh/serve-cli": "workspace:*",
+    "@graphql-mesh/utils": "workspace:*",
+    "@omnigraph/odata": "workspace:*",
+    "graphql": "^16.9.0"
+  }
+}
