+++ packages/loaders/odata/src/resolvers/rootResolver.ts
@@ -1,0 +1,5 @@
+import type { GraphQLFieldResolver } from 'graphql';
+
+export const rootResolver: GraphQLFieldResolver<any, any> = function rootResolver(root) {
+  return root;
+};
