+++ packages/loaders/odata/src/utils/QueryOptionsFields.ts
@@ -1,0 +1,30 @@
+export const QUERY_OPTIONS_FIELDS = {
+  orderby: {
+    type: 'String',
+    description:
+      'A data service URI with a $orderby System Query Option specifies an expression for determining what values are used to order the collection of Entries identified by the Resource Path section of the URI. This query option is only supported when the resource path identifies a Collection of Entries.',
+  },
+  top: {
+    type: 'Int',
+    description:
+      'A data service URI with a $top System Query Option identifies a subset of the Entries in the Collection of Entries identified by the Resource Path section of the URI. This subset is formed by selecting only the first N items of the set, where N is an integer greater than or equal to zero specified by this query option. If a value less than zero is specified, the URI should be considered malformed.',
+  },
+  skip: {
+    type: 'Int',
+    description:
+      'A data service URI with a $skip System Query Option identifies a subset of the Entries in the Collection of Entries identified by the Resource Path section of the URI. That subset is defined by seeking N Entries into the Collection and selecting only the remaining Entries (starting with Entry N+1). N is an integer greater than or equal to zero specified by this query option. If a value less than zero is specified, the URI should be considered malformed.',
+  },
+  filter: {
+    type: 'String',
+    description:
+      'A URI with a $filter System Query Option identifies a subset of the Entries from the Collection of Entries identified by the Resource Path section of the URI. The subset is determined by selecting only the Entries that satisfy the predicate expression specified by the query option.',
+  },
+  inlinecount: {
+    type: 'InlineCount',
+    description:
+      'A URI with a $inlinecount System Query Option specifies that the response to the request includes a count of the number of Entries in the Collection of Entries identified by the Resource Path section of the URI. The count must be calculated after applying any $filter System Query Options present in the URI. The set of valid values for the $inlinecount query option are shown in the table below. If a value other than one shown in Table 4 is specified the URI is considered malformed.',
+  },
+  count: {
+    type: 'Boolean',
+  },
+};
