+++ packages/loaders/odata/src/utils/addIdentifierToUrl.ts
@@ -1,0 +1,8 @@
+export function addIdentifierToUrl(
+  url: URL,
+  identifierFieldName: string,
+  identifierFieldTypeRef: string,
+  args: any,
+) {
+  url.href += `/${args[identifierFieldName]}/`;
+}
