+++ packages/loaders/odata/src/utils/getUrlString.ts
@@ -1,0 +1,3 @@
+export function getUrlString(url: URL) {
+  return decodeURIComponent(url.toString()).split('+').join(' ');
+}
