+++ packages/transports/odata/src/index.ts
@@ -1,0 +1,13 @@
+import { createDefaultExecutor, type Transport } from '@graphql-mesh/transport-common';
+import { processDirectives } from '@omnigraph/odata';
+
+export default {
+  getSubgraphExecutor({ subgraph, fetch }) {
+    return createDefaultExecutor(
+      processDirectives({
+        schema: subgraph,
+        fetchFn: fetch,
+      }),
+    );
+  },
+} satisfies Transport;
